:white_check_mark: Move fixture in conftest file
