+++ tests/conftest.py
@@ -2,6 +2,8 @@
 
 import pytest
 
+from jpyltime.utils import Attribute
+
 
 @pytest.fixture
 def client():
@@ -11,3 +13,15 @@ def client():
     from jpyltime.app import app
 
     return TestClient(app)
+
+
+@pytest.fixture
+def attributes():
+    attributes = [
+        Attribute(official_name="First name", custom_name="Prénom", anonymize=False),
+        Attribute(official_name="Weight", custom_name="Poids", anonymize=False),
+        Attribute(official_name="Medication", custom_name="Médicaments", anonymize=False),
+        Attribute(official_name="Birthdate", custom_name="Anniversaire", anonymize=True),
+    ]
+    requested_attributes = {attribute.official_name: attribute for attribute in attributes}
+    return requested_attributes
