+++ tests/postprocessing_test.py
@@ -42,18 +42,6 @@ def example_dataframe():
     return df
 
 
-@pytest.fixture
-def attributes():
-    attributes = [
-        Attribute(official_name="First name", custom_name="Prénom", anonymize=False),
-        Attribute(official_name="Weight", custom_name="Poids", anonymize=False),
-        Attribute(official_name="Medication", custom_name="Médicaments", anonymize=False),
-        Attribute(official_name="Birthdate", custom_name="Anniversaire", anonymize=True),
-    ]
-    requested_attributes = {attribute.official_name: attribute for attribute in attributes}
-    return requested_attributes
-
-
 def test_postprocessing(map_attributes, example_dataframe, attributes):
     anonymization_symbol = "*"
     postprocessing = FHIR2DS_Postprocessing(
